schedtime: Wait for tasks to finish when hitting Ctrl-C
There's been a long-standing bug in this script where we can try
collecting task lifetimes even though they haven't died properly and
passed through the kernel's exit_thread() function.

It turns out this is because we were skipping the proc.wait() call
if the user interrupts execution, thereby not properly waiting for all
tasks to finish.

Reported-by: @pkolaczk